requirements.txt
flask
flask_wtf
wtforms
maxminddb-geolite2
